requirements.txt
appdirs==1.4.4
backcall==0.2.0
beautifulsoup4==4.11.2
blinker==1.5
Brotli==1.0.9
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==39.0.2
dataclasses==0.6
decorator==5.1.1
fs==2.4.16
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.56.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httplib2==0.21.0
hyperframe==6.0.1
idna==3.4
importlib-resources==5.4.0
jedi==0.17.2
kaitaistruct==0.10
lz4==3.1.10
oauthlib==3.2.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pyfatfs==1.0.5
Pygments==2.14.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
PyVirtualDisplay==3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
samloader @ git+https://github.com/nlscc/samloader.git@95d2ac8fb9027b7908d201e4ce807a5b338f923a
selenium==3.141.0
selenium-wire==5.0.0
six==1.16.0
soupsieve==2.3.2.post1
tomli==1.2.3
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
wsproto==1.0.0
zipp==3.6.0
zstandard==0.20.0
bs4
construct 
tqdm
ipdb
ipython
ipython-genutils
